Move TypeCast to Basic.py
  Inserting access declaration to all dimensions of ops_dat.
  Removing TypeCast from parameters.
