+++ devito/ops/types.py
@@ -3,7 +3,7 @@ import sympy
 
 import devito.types.basic as basic
 
-from devito.tools import ctypes_to_cstr, dtype_to_cstr
+from devito.tools import dtype_to_cstr
 from devito.ops.utils import namespace
 
 
@@ -147,18 +147,3 @@ class OpsDat(basic.LocalObject):
     @property
     def _C_typename(self):
         return namespace['ops_dat_type']
-
-
-class TypeCast(basic.Object):
-    """
-        A node encapsulating a cast to a raw C type.
-    """
-
-    def __init__(self, name, dtype, cast_to_ctype):
-        self.cast_to_ctype = cast_to_ctype
-        self.dtype = dtype
-        self.name = str(name)
-
-    @property
-    def _C_name(self):
-        return '(%s)%s' % (ctypes_to_cstr(self.cast_to_ctype), str(self.name))
